Creating cloud formation template for the ECS service
[Problem]
To make the project more deployable for other organizations and resilient, the SMS alert system is migrating to using
Docker deployments on top of ECS. Instead of configuring everything by hand, which is error prone during set-up, making
changes, and makes it difficult for other orgs to replicate, we opted to set up the entire stack using CloudFormation.

[Solution]
Create a stack using cloudformation. A Go DSL was used for type checking and making it possible to easily create Gamma
and Prod stacks.

Heavily influenced by:
http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/quickref-ecs.html

[Test]
Created the Gamma stack